requirements.txt
dependency-injector
flask
flask-marshmallow
typing_extensions
marshmallow_generic